backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.3.0

# Google Gemini AI
google-generativeai==0.3.1

# Audio/Video Processing
moviepy==1.0.3
pydub==0.25.1
SpeechRecognition==3.10.0

# Utilities
aiofiles==23.2.1
httpx==0.25.2
Pillow==10.1.0